% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pricingKernel.R
\name{A}
\alias{A}
\title{An (Pricing kernel helper)}
\usage{
A(An, Bnt, SigmaL0, Sigma)
}
\arguments{
\item{An}{previous value of An (i.e. An-1)}

\item{Bnt}{previous value of Bn' (i.e. Bn-1')}

\item{SigmaL0}{pricing kernel parameter estimated in VAR Calibration (Step 2)}

\item{Sigma}{covariance matrix from OLS}
}
\value{
An
}
\description{
Log price of a bond is An + Bn' zt where zt is VAR state vector (Step 1)
}

