% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{accountValueWF}
\alias{accountValueWF}
\title{Account Value without floor on return}
\usage{
accountValueWF(employee, surviveInfo.out, assumptions = planVariables,
  oneyear = term1, tenyear = term10)
}
\arguments{
\item{employee}{A row from the demoInfo dataset}

\item{surviveInfo.out}{output of surviveInfo()}

\item{assumptions}{Plan features and assumptions - by default
the variables in planVariables are used.}

\item{oneyear}{1 year treasury bond yields. By default, deterministic scenario
output is used.}

\item{tenyear}{10 year treasury bond yields. By default, deterministic scenario
output is used.}

\item{r}{The maximum retirement age}
}
\value{
A data frame containing the account value boy, account value
 after growth, salary, contribution (eoy), account value (eoy),
 benefit payments (boy) and discount benefit payment for each year
 until retirement
}
\description{
Account Value without floor on return
}
\details{
WF stands for with floor. For the most part, this function
 is very similar to accountValueNF()
}
\examples{
surviveInfo.out <- surviveInfo(demoInfo[5,]);
 accountValueWF.out <- accountValueWF(demoInfo[5,], surviveInfo.out)
}

