% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pricingKernel.R
\name{B}
\alias{B}
\title{Bn' (Pricing kernel helper)}
\usage{
B(Bnt, Phi, delta1, SigmaL1)
}
\arguments{
\item{Bnt}{previous value of Bn' (i.e. Bn-1')}

\item{Phi}{coefficients from OLS}

\item{delta1}{vector of constants}
}
\value{
Bn'
}
\description{
Log price of a bond is An + Bn' zt where zt is VAR state vector (Step 1)
}

