% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{bmarkScenarios}
\alias{bmarkScenarios}
\title{Read SOA stochastic scenarios}
\usage{
bmarkScenarios(file)
}
\arguments{
\item{file}{The file path to the CSV file containing the stochastic scenarios}
}
\value{
A list containing data frames for the term1 and term10 interest rates
}
\description{
Imports the SOA stochastic scenarios into a list structure
 containing data frames for term1 and term10 interest rates
}
\details{
file argument depends on where the CSV  file is stored
}
\examples{
soaScenarios <- bmarkScenarios("~/Desktop/StocVal/Data/soascenario.csv")
}

