% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riskNeutral.R
\name{CBP3CFV}
\alias{CBP3CFV}
\title{Helper function for floorOptionPBO}
\usage{
CBP3CFV(x, r, survFactor, discFactor, bondIndex, av, vol, floor)
}
\arguments{
\item{x}{Actual age}

\item{r}{Retirement age}

\item{survFactor}{Survival factor for the current year}

\item{discFactor}{Discount factor for the current year}

\item{bondIndex}{The bond index for the current year}

\item{av}{The account value at the beginning of the year}

\item{vol}{The volatility on a long term treasury bond}

\item{floor}{The floor rate}
}
\value{
The PBO for the given input parameters
}
\description{
Used to calculate PBO for closed form valuation
}

