% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{cholesky}
\alias{cholesky}
\title{Cholesky Decomposition}
\usage{
cholesky(cholinput)
}
\arguments{
\item{cholinput}{a 4x4 correlation matrix for the short rate, credit spread,
excess return and inflation rate}
}
\value{
a 4x4 matrix representing the Cholesky decomposition
}
\description{
Computes the Cholesky Decomposition of a correlation matrix
}
\examples{
cholesky.out <- cholesky(cholinput)
}

