% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholinput}
\alias{cholinput}
\title{DATA: cholesky correlation matrix}
\format{4x4 matrix}
\usage{
cholinput
}
\description{
Contains 4x4 correlation matrix for short rate, credit spread,
 excess return and inflation rate
}
\keyword{datasets}

