% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pricingKernel.R
\name{compare_yield_fit}
\alias{compare_yield_fit}
\title{Compare fitted yields to historical yields (PLOTS)}
\usage{
compare_yield_fit(varData, histYields, calibrate_VAR.out, Phi, Sigma, N = c(2,
  3, 5, 10))
}
\arguments{
\item{varData}{VAR columns used in OLS (i.e. onemonth, inflation, tenyear, stock)}

\item{histYields}{histYields to calibrate term structure with (i.e. twoyear, threeyear,
fiveyear, tenyear)}

\item{calibrate_VAR.out}{output of calibrate_VAR}

\item{Phi}{coefficients of OLS}

\item{Sigma}{covariance matrix of OLS}

\item{N}{specifies the terms used in histYields (i.e. 2, 3, 5, 10).
Should be the same as parameter than was passed to calibrate_VAR().}
}
\value{
see plots
}
\description{
Plots historical vs fitted yields. Note that this function
 WILL work for any number of VAR components as long as calibrate_VAR.out
 input is correct (however currently calibrate_VAR() only takes 4 VAR components).
}
\details{
For example,
 marketData <- readRDS("~/Dropbox/Research/StocVal/data/Canada/varinput_canada.Rda");
 varData <- data.frame(onemonth=marketData$onemonth, inflation=marketData$inflation,
   tenyear=marketData$tenyear, stock=marketData$stock);
 histYields <- data.frame(twoyear=marketData$twoyear, threeyear=marketData$threeyear,
   fiveyear=marketData$fiveyear, tenyear=marketData$tenyear);
 N <- c(2,3,5,10)
 calibrate_VAR.out <- calibrate_VAR(varData, histYields, Phi, Sigma, N);
 compare_yield_fit(varData, histYields, calibrate_VAR.out, Phi, Sigma, N)
}

