% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crspinput}
\alias{crspinput}
\title{DATA: Credit Rating Spread Input parameters}
\format{1D vector with 5 entries}
\usage{
crspinput
}
\description{
1D vector containing b, tau, scenarios, projyears,
 seed for hull white short rates scenario generation
}
\keyword{datasets}

