% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{cubic_spline}
\alias{cubic_spline}
\title{Cubic Spline}
\usage{
cubic_spline(termStruct)
}
\arguments{
\item{termStruct}{Data frame containing term and USD risk free rate. Maximum term is 30.}
}
\value{
A data frame containing the interpolated USD risk free rates.
}
\description{
Computes the cubic spline linear interpolation for unknown yield terms
}
\details{
This function must be called to perform the stochastic valuation
}
\examples{
termStruct.out <- cubic_spline(termStruct)
}

