% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{curve_extensionCF}
\alias{curve_extensionCF}
\title{Curve Extension (Swap, constant forward)}
\usage{
curve_extensionCF(termStruct.out)
}
\arguments{
\item{termStruct.out}{Contains the linearly interpolated risk free rates
from cubic_spline()}
}
\value{
The curve extension for an additional 70 years (effective annual zero rate)
}
\description{
Calculates the curve extension until time 100 using swap (constant forward)
}
\examples{
termStruct.out <- cubic_spline(termStruct)
 termStruct.extension <- curve_extensionCF(termStruct.out)
}

