% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoInfo}
\alias{demoInfo}
\title{DATA: Demographic Info for Plan}
\format{A data frame with 5 rows and 7 columns}
\usage{
demoInfo
}
\description{
Employee salaries, ages and account values
}
\keyword{datasets}

