% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{determScenario}
\alias{determScenario}
\title{Generate a deterministic scenario}
\usage{
determScenario(termStruct.out, params = NS_params)
}
\arguments{
\item{termStruct.out}{output of cubic_spline()}

\item{params}{Neilson and Siegel parameters for calculating the curve extension}
}
\value{
A list containing interest rates for various terms, inflation rate,
 credit rate, total yield, equity index and bond index
}
\description{
Generates deterministic scenario for variables
 interest rate term, inflation rate, credit spread, total yield, equity
 index, and bond index
}
\details{
output is similar to stochastic scenarios except that
 only one scenario is returned that the scenario is not random.
 Note that time goes from 0 to 99 for this scenario.
}
\examples{
termStruct.out <- cubic_spline(termStruct)
 determScenario.out <- determScenario(termStruct.out)
}

