% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ers}
\alias{ers}
\title{Equity return scenario generation}
\usage{
ers(rpi, volterm, termStruct.extension)
}
\arguments{
\item{rpi}{Equity-risk premium input parameters}

\item{volterm}{Volatility for bonds with terms 1 to 100 length inclusive (equity)}

\item{termStruct.extension}{output of curve_extensionNS()}
}
\value{
The equity excess return for each scenario
}
\description{
Generates stochastic scenarios for the excess equity return
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 ers.out <- ers(rpi, volterm, termStruct.extension)
}

