% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ersTotal}
\alias{ersTotal}
\title{Total equity return for scenarios}
\usage{
ersTotal(hw.out, ers.out)
}
\arguments{
\item{hw.out}{output of hull_white()}

\item{ers.out}{output of ers()}
}
\value{
A data frame containing the equity total return for each scenario
}
\description{
Derives the equity total return from the excess equity return.
 Note this method simply adds two data frames.
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 ers.out <- ers(rpi, volterm, termStruct.extension)
 ersTotal.out <- ersTotal(ers.out)
}

