% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{esga}
\alias{esga}
\title{ESGA}
\usage{
esga(esgin, credit_spread, volterm.equity, inflation_mean, volterm.inflation,
  cholesky.out, termStruct.extension)
}
\arguments{
\item{esgin}{ESG scenario generation input}

\item{credit_spread}{Volatility for credit spreads}

\item{volterm.equity}{Volatility for bonds with terms 1 to 100 length inclusive}

\item{cholesky.out}{output of cholesky()}

\item{termStruct.extension}{output of curve_extensionNS()}

\item{inflation_in}{Inputs for inflation scenario generation}
}
\value{
A list of data frames with values for scenario 1 to 1000 and terms 0 to 99
}
\description{
Holistic ESG for short rate, credit spread, equity
             return and inflation with correlation
}
\details{
List indexes: Index 1 is short rate, Index 2 is credit spread, Index 3
 is total equity return and Index 4 is inflation rate. An important feature
 of this function is that it uses the Cholesky decomposition of the
 variance-covariance matrix between these varaibles.
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 credit_spread <- rep(0.01,100);
 cholesky.out <- cholesky(cholinput);
 esga.out <- esga(esgin, credit_spread, volterm.equity, inflation_mean,
 volterm.inflation , cholesky.out, termStruct.extension)
}

