% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esgin}
\alias{esgin}
\title{DATA: ESG scenario generation input}
\format{1D vector with 8 elements}
\usage{
esgin
}
\description{
Input parameters for ESG scenario generation containing
 Interest Rate - b, Interest Rate - tau, Equity Risk premium (0 since risk neutral),
 Credit Spread - b, Credit Spread - tau, # of scenarios, projection years,
 random number seed
}
\details{
Interest Rate - b is from srinput;
 Interest Rate - tau is from srinput;
 Equity Risk Premium is from rpi;
 Credit Spread - b is from crspinput;
 Credit Spread - tau is from crspinput;
 # of scenarios is from srinput;
 projection years is from srinput;
 Random number seed is from srinput
}
\keyword{datasets}

