% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riskNeutral.R
\name{eur_call_bsm}
\alias{eur_call_bsm}
\title{Cost of European Call Option}
\usage{
eur_call_bsm(S, K, r, sigma, d, term)
}
\arguments{
\item{S}{The current price of the underlying asset (i.e. stock)}

\item{K}{The strike / exercise price of the option}

\item{r}{The risk free interest rate}

\item{sigma}{The volatility of the option}

\item{d}{The continuously compounded dividend rate}

\item{term}{The time to expiration}
}
\value{
The cost of a European call option
}
\description{
Calculates the cost of a European call option under the
 black scholes model.
}
\details{
bsm stands for the black scholes model.
}
\examples{
eur_call_bsm(40,40,0.05,0.20,0.02,1)
}

