% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{floorOption}
\alias{floorOption}
\title{Cost of floor option}
\usage{
floorOption(accountValueNF.out, accountValueWF.out, oneyear = term1,
  tenyear = term10)
}
\arguments{
\item{accountValueNF.out}{output of accountValueNF()}

\item{accountValueWF.out}{output of accountValueWF()}

\item{oneyear}{1 year treasury bond yields. By default, deterministic scenario
output is used.}

\item{tenyear}{10 year treasury bond yields. By default, deterministic scenario
output is used.}
}
\value{
The cost of the embedded option
}
\description{
Returns the cost of the embedded floor option, calculated
 by taking the difference in the payment with the option vs
 without the option.
}
\details{
The option is a floor on the minimum return, so it increases
 the value of the benefit for the employee.
}
\examples{
surviveInfo.out <- surviveInfo(demoInfo[5,]);
 accountValueNF.out <- accountValueNF(demoInfo[5,], surviveInfo.out);
 accountValueWF.out <- accountValueWF(demoInfo[5,], surviveInfo.out);
 floorOption.cost <- floorOption(accountValueNF.out, accountValueWF.out)
}

