% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{floorOptionPBO}
\alias{floorOptionPBO}
\title{Cost of floor option (pbo)}
\usage{
floorOptionPBO(employee, accountValueNF.out, accountValueWF.out,
  surviveInfo.out, assumptions = planVariables, oneyear = term1,
  tenyear = term10)
}
\arguments{
\item{employee}{A row from the demoInfo dataset}

\item{accountValueNF.out}{output of accountValueNF()}

\item{accountValueWF.out}{output of accountValueWF()}

\item{surviveInfo.out}{output of surviveInfo()}

\item{assumptions}{Plan features and assumptions - by default
the variables in planVariables are used.}

\item{oneyear}{1 year treasury bond yields. By default, deterministic scenario
output is used.}

\item{tenyear}{10 year treasury bond yields. By default, deterministic scenario
output is used.}
}
\value{
The cost of the embedded option
}
\description{
Cost of floor option (pbo)
}
\details{
The option cost will be different when a stochastic valuation is
 used. This is a different way of calculating the cost of the floor option.
}
\examples{
surviveInfo.out <- surviveInfo(demoInfo[5,]);
 accountValueNF.out <- accountValueNF(demoInfo[5,], surviveInfo.out);
 accountValueWF.out <- accountValueWF(demoInfo[5,], surviveInfo.out);
 floorOptionPBO.cost <- floorOptionPBO(demoInfo[5,], accountValueNF.out,
 accountValueWF.out, surviveInfo.out)
}

