% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{hull_white}
\alias{hull_white}
\title{1-factor Hull White}
\usage{
hull_white(srinput, termStruct.extension, termofyield = 1)
}
\arguments{
\item{srinput}{Hull White model parameters. c(b, tau, scen, prjy, rns)}

\item{termStruct.extension}{output of curve_extensionNS()}

\item{termofyield}{should be 1 or function won't work}
}
\value{
Stochastic short rate scenarios
}
\description{
Short rate interest model used to generate stochastic scenarios
}
\details{
In srinput, scen represents the number of scenarios to generate,
 prjy represents the number of projected years, and rns represents the
 random number seed. It should also be pointed out that hull_white() can
 also be used to perform credit spread scenario generation not just short rate
 scenario generation (see interestModelsDemo)
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 hw.out <- hull_white(srinput, termStruct.extension, termofyield=1)
}

