% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{hull_whiteBond}
\alias{hull_whiteBond}
\title{1-factor Hull White (bond prices)}
\usage{
hull_whiteBond(hw.out)
}
\arguments{
\item{hw.out}{Output of hull_white() containing short rate scenarios}
}
\value{
Bond prices corresponding to hull white short rate scenarios
}
\description{
Transform hull white short rates into bond prices -> P = 1/(1+r)
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 hw.out <- hull_white(srinput, termStruct.extension, termofyield=1);
 hw.bond <- hull_whiteBond(hw.out)
}

