% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{hw_theta}
\alias{hw_theta}
\title{1-factor Hull White Theta}
\usage{
hw_theta(srinput, termStruct.extension)
}
\arguments{
\item{srinput}{Hull White model parameters. c(b, tau, scen, prjy, rns)}

\item{termStruct.extension}{output of curve_extensionNS()}
}
\value{
A theta vector
}
\description{
Calculates theta for 1-factor hull white short rate model
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 hw.theta <- hw_theta(srinput, termStruct.extension)
}

