% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation_in}
\alias{inflation_in}
\title{DATA: Inputs for inflation scenario generation}
\format{1D vector with 3 entries}
\usage{
inflation_in
}
\description{
1D vector containing # of scenarios, projection years,
 and random number seed
}
\keyword{datasets}

