% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation_mean}
\alias{inflation_mean}
\title{DATA: Inflation mean by term}
\format{1D vector with 100 etnries}
\usage{
inflation_mean
}
\description{
1D vector containing expected inflation for each term
}
\keyword{datasets}

