% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{inflationrs}
\alias{inflationrs}
\title{Inflation scenario generation}
\usage{
inflationrs(inflation_in, inflation_mean, volatility.inflation)
}
\arguments{
\item{inflation_in}{Parameters for the simulation scenario generation}

\item{inflation_mean}{Inflation mean for each term}

\item{volatility.inflation}{Implied volatility term structure}
}
\value{
A data frame containing inflation scenarios for each term
}
\description{
Generate inflation scenarios using calibrated inflation mean and
 volatility information
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 inflationrs.out <- inflationrs(inflation_in, inflation_mean,
   volatility.inflation)
}

