% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{meanScenarios}
\alias{meanScenarios}
\title{Calculate the average stochastic scenario}
\usage{
meanScenarios(scenarios)
}
\arguments{
\item{scenarios}{A data frame containing the scenarios, in format of the output
of hull_white()}
}
\value{
A vector containing the mean of the stochastic scenarios at each
 time
}
\description{
Calculates the mean of the stochastic scenarios at each time
}
\examples{
soaScenarios <- bmarkScenarios("~/Desktop/StocVal/Data/soascenario.csv");
 meanScenarios.out <- meanScenarios(soaScenarios$term1)
}

