% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortalityInfo}
\alias{mortalityInfo}
\title{DATA: Mortality Info for Plan}
\format{A data frame with 111 rows and 3 columns}
\usage{
mortalityInfo
}
\description{
qxm and qxf for age 0 to 110 inclusive
}
\keyword{datasets}

