% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NS_params}
\alias{NS_params}
\title{DATA: NS_params}
\format{1D vector with 6 entries}
\usage{
NS_params
}
\description{
Parameters for the Neilson and Siegel extrapolation containing
 m, B0, B1, B2, tau1, tau2
}
\details{
The first entry (m) represents the time to maturity
}
\keyword{datasets}

