% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{nyearZero}
\alias{nyearZero}
\title{N-year zero rates}
\usage{
nyearZero(hw.out, termStruct.extension, srinput, term = 10)
}
\arguments{
\item{hw.out}{output of hull_white()}

\item{termStruct.extension}{output of curve_extensionNS()}

\item{srinput}{Hull-white model parameters}
}
\value{
A data frame containing n-year zero rates for each scenario
}
\description{
Derives N-year zero rates from 1-factor hull white
}
\details{
Due to the way in which theta was calculated, the last three columns
 of hw.out are not accurate. Therefore, the last four columns of nyearZero(..., term=2)
 are not accurate, and so on. Also note that ,
 nyearZero(,...term=1) and hw.out are equivalent
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 hw.out <- hull_white(srinput, termStruct.extension, termofyield=1);
 nyearZero.out <- nyearZero(hw.out, termStruct.extension, srinput, term=10)
}

