% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{parYield}
\alias{parYield}
\title{Par Yield}
\usage{
parYield(hw.out, termStruct.extension, srinput, term = 5)
}
\arguments{
\item{hw.out}{output of hull_white()}

\item{termStruct.extension}{output of curve_extensionNS()}

\item{srinput}{Hull-white model parameters}

\item{term}{The term of the par yield}
}
\value{
A data frame containing par yield rates for each scenario
}
\description{
Calculates par yield using results of Hull White scenario generation
}
\details{
hw.out uses a term of 10 years. This function allows for a different
 term (i.e. 5 years). However, it has three nested loops, so can take a while.
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 hw.out <- hull_white(srinput, termStruct.extension, termofyield=1);
 parYield.out <- parYield(hw.out, termStruct.extension, srinput, term=5)
}

