% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{planSummaryDeterm}
\alias{planSummaryDeterm}
\title{Embedded option prices for each employee (Deterministic)}
\usage{
planSummaryDeterm(demoInfo, assumptions = planVariables)
}
\arguments{
\item{demoInfo}{Demographic info for plan}

\item{assumptions}{Plan features and assumptions - by default
the variables in planVariables are used.}
}
\value{
Option costs for each employee
}
\description{
Calculates the embedded option prices for each employee using the deterministic
 scenario output from floorOption() and floorOptionPBO()
}
\examples{
planSummaryDeterm.out <- planSummaryDeterm(demoInfo,planVariables)
}

