% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{planSummaryStoc}
\alias{planSummaryStoc}
\title{Embedded option prices for each employee (Stochastic)}
\usage{
planSummaryStoc(demoInfo, stochasticScenarios.out,
  assumptions = planVariables)
}
\arguments{
\item{demoInfo}{Demographic info for plan}

\item{stochasticScenarios.out}{output of stochasticScenarios()}

\item{assumptions}{Plan features and assumptions - by default
the variables in planVariables are used.}
}
\value{
A list containing the total floor option cost for each scenario,
 the total floor option (PBO) cost for each scenario and a data frame
 containing the expected value and mean costs for each scenario
}
\description{
Calculates the embedded option prices for each employee using
 the stochastic scenario output from stocfloorOption() and stocfloorOptionPBO()
}
\examples{
planSummaryStoc.out <- planSummaryStoc(demoInfo, stochasticScenarios.out,
 planVariables)
}

