% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planVariables}
\alias{planVariables}
\title{DATA: Plan Features and Assumptions}
\format{List with 5 entries: expected salary growth rate,
 contribution rate, vesting period, cash balance plan floor,
 volatility on treasury bond, maximum retirement age}
\usage{
planVariables
}
\description{
DB plan parameters used for modeling
}
\keyword{datasets}

