% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpi}
\alias{rpi}
\title{DATA: Excess Equity Input parameters}
\format{1D vector with 4 entries}
\usage{
rpi
}
\description{
1D vector containing risk premium (0 since risk neutral),
 # of scenarios, projection years, random number seed
}
\keyword{datasets}

