% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma}
\alias{Sigma}
\title{DATA: Sigma (VAR with 4 components)}
\format{A 4x4 data frame (should be converted to matrix)}
\usage{
Sigma
}
\description{
Covariance parameter for VAR model
}
\keyword{datasets}

