% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srinput}
\alias{srinput}
\title{DATA: Hull White Input parameters}
\format{1D vector with 5 entries}
\usage{
srinput
}
\description{
1D vector containing b, tau, scenarios, projyears,
 seed for hull white short rate scenario generation
}
\keyword{datasets}

