% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{stocfloorOptionPBO}
\alias{stocfloorOptionPBO}
\title{Stochastic cost of floor option (pbo)}
\usage{
stocfloorOptionPBO(employee, surviveInfo.out, stochasticScenarios.out,
  assumptions = planVariables)
}
\arguments{
\item{employee}{A row from the demoInfo dataset}

\item{surviveInfo.out}{output of surviveInfo()}

\item{stochasticScenarios.out}{output of stochasticScenarios()}
}
\value{
A list containing the cost of the option under each scenario, mean
 and variance.
}
\description{
Returns a list containing the costs of the floor option
 under each stochastic scenario, the expected value and variance
}
\examples{
surviveInfo.out <- surviveInfo(demoInfo[5,]);
 termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 credit_spread <- rep(0.01,100);
 hw.out <- hull_white(srinput, termStruct.extension, termofyield=1);
 cholesky.out <- cholesky(cholinput);
 esga.out <- esga(esgin, credit_spread, volterm.equity, inflation_mean,
 volterm.inflation , cholesky.out, termStruct.extension);
 stochasticScenarios.out <- stochasticScenarios(hw.out, termStruct.extension,
 srinput, esga.out)
 stocfloorOptionPBO.cost <- stocfloorOptionPBO(demoInfo[5,], surviveInfo.out,
 stochasticScenarios.out)
}

