% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{stochasticScenarios}
\alias{stochasticScenarios}
\title{Generate stochastic scenarios}
\usage{
stochasticScenarios(hw.out, termStruct.extension, srinput, esga.out)
}
\arguments{
\item{hw.out}{output of hull_white()}

\item{termStruct.extension}{output of curve_extensionNS()}

\item{srinput}{hull white input parameters}

\item{esga.out}{output of esga()}
}
\value{
A list containing interest rates for various terms, inflation rate,
 credit rate, total yield, equity index and bond index
}
\description{
Convert esga() output into a form that can be read by
 the valuation functions
}
\examples{
termStruct.out <- cubic_spline(termStruct);
 termStruct.extension <- curve_extensionNS(termStruct.out);
 credit_spread <- rep(0.01,100);
 hw.out <- hull_white(srinput, termStruct.extension, termofyield=1);
 cholesky.out <- cholesky(cholinput);
 esga.out <- esga(esgin, credit_spread, volterm.equity, inflation_mean,
 volterm.inflation , cholesky.out, termStruct.extension);
 stochasticScenarios.out <- stochasticScenarios(hw.out, termStruct.extension,
 srinput, esga.out)
}

