% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{surviveInfo}
\alias{surviveInfo}
\title{Survivorship Information for employee}
\usage{
surviveInfo(employee, assumptions = planVariables,
  mortTable = mortalityInfo)
}
\arguments{
\item{employee}{A row from the demoInfo dataset}

\item{assumptions}{Plan features and assumptions - by default
the variables in planVariables are used.}

\item{mortTable}{The mortality table to use containing qxm and qxf
columns for age 0 to 110. By default the columns in the mortalityInfo
dataset are used.}
}
\value{
The mortality rate, termination rate, probability of decrement
 during year and survival factor at the beginning of the year for
 the employee.
}
\description{
Creates a data frame containing the decrements for the
 employee.
}
\details{
The mortality table used can be found in the mortalityInfo
 dataset.
}
\examples{
surviveInfo.out <- surviveInfo(demoInfo[5,])
}

