% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term1}
\alias{term1}
\title{DATA: 1-year deterministic yields}
\format{1D vector with 100 entries corresponding to time 0-99}
\usage{
term1
}
\description{
output of determScenario() -> term1
}
\keyword{datasets}

