% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term10}
\alias{term10}
\title{DATA: 10-year deterministic yields}
\format{1D vector with 100 entries corresponding to time 0-99}
\usage{
term10
}
\description{
output of determScenario() -> term10
}
\keyword{datasets}

