% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuation.R
\name{termAssumption}
\alias{termAssumption}
\title{Termination Assumption for an employee}
\usage{
termAssumption(employee, r = 65)
}
\arguments{
\item{employee}{A row from the demoInfo dataset}

\item{r}{The maximum retirement age}
}
\value{
A data frame containing the termination assumption
 for the employee from age x to age r inclusive
}
\description{
Termination Assumption for an employee
}
\examples{
employee <- demoInfo[5,];
 termAssumption.out <- termAssumption(employee)
}

