% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{termStruct}
\alias{termStruct}
\title{DATA: Term Structure rates}
\format{A data frame with 12 rows and 2 variables}
\usage{
termStruct
}
\description{
Risk free interest rates
}
\keyword{datasets}

