% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volterm.equity}
\alias{volterm.equity}
\title{DATA: Volatility by term of bond}
\format{1D vector with 100 entries}
\usage{
volterm.equity
}
\description{
Volatility for bonds with terms 1 to 100 length inclusive
}
\keyword{datasets}

