% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volterm.inflation}
\alias{volterm.inflation}
\title{DATA: Volatility by term from inflation}
\format{1D vector with 100 entries}
\usage{
volterm.inflation
}
\description{
Volatility for inflation for term 1 to 100 inclusive
}
\keyword{datasets}

